+++ src/components/Button/IconButtonComponent.jsx
@@ -5,11 +5,12 @@ import React from 'react';
  *
  * @param {React.ReactNode} icon - The icon element to display.
  * @param {string} label - The label or text associated with the button.
+ * @param {string} isSubmitBtn - The label or text associated with the button.
  * @param {function} onClick - The click event handler for the button.
  */
-const IconButton = ({ icon, label, onClick, className }) => {
+const IconButton = ({ icon, label, onClick, className , isSubmitBtn}) => {
   return (
-    <button type="button" className={`btn  ${className ? className : "bg-primary text-white"} `} onClick={onClick}>
+    <button type={isSubmitBtn ? "submit" : "button"} className={`btn  ${className ? className : "bg-primary text-white"} `} onClick={onClick}>
       <span className='ml-2 mr-2'>{label}</span>
       <span className="text-xl">{icon}</span>
