+++ src/components/Input/TextField/DefaultInput.jsx
@@ -10,6 +10,7 @@ import LoadingSvg from 'src/assets/svg/LoadingSvg';
  * @param {function} onChange - The callback function triggered on input change.
  * @param {boolean} showPassword - Whether to show the password field.
  * @param {boolean} icon - Whether to show the icon or not
+ * @param {boolean} secondIcon - Whether to show the second icon or not
  * @param {function} onClick - The callback function triggered when the arrow icon is clicked (optional).
  * @param {boolean} loading - Whether to show the loading SVG.
  * @param {string} type - The type of input field (password, email, text).
@@ -24,7 +25,9 @@ const DefaultInput = ({
   loading = false,
   type = 'text',
   error,
-  icon
+  icon,
+  secondIcon,
+  onClickSecondIcon
 }) => {
 
   return (
@@ -39,22 +42,19 @@ const DefaultInput = ({
               className={`absolute inset-y-0 left-2 flex items-center pr-3 ${error ? 'pt-4' : 'pt-10'
                 } `}
             >
-              {/* <svg
-                onClick={() => value && onClick && onClick(value)}
-                className={`w-8 h-13 text-gray-600 ${!hasValue ? 'text-gray-400' : 'text-gray-900 hover:text-gray-700'
-                  }`}
-                viewBox="0 0 20 20"
-                fill="currentColor"
-              >
-                <path
-                  fillRule="evenodd"
-                  d="M10 3a1 1 0 0 1 .707.293l4 4a1 1 0 0 1 0 1.414l-4 4a1 1 0 0 1-1.414-1.414L11.586 9H3a1 1 0 0 1 0-2h8.586l-1.293-1.293A1 1 0 0 1 10 3z"
-                  clipRule="evenodd"
-                />
-              </svg> */}
               {icon}
             </div>
           )
+        )
+      }
+      {
+        secondIcon && (
+            <div
+              className={`absolute inset-y-0 right-2 flex items-center pr-3 ${error ? 'pt-4' : 'pt-10'
+                } `}
+            >
+             <button onClick={() => onClickSecondIcon && onClickSecondIcon()} > {secondIcon}</button>
+            </div>
         )
       }
       <input
